fix: Recursive parsing of attributes
The `test_parse` macro is changed to recursively parse attributes,
filtering out the `#[ignore]` macro. This should also make it fairly
trivial to add additional macros, such as `#[should_ignore]`.

There is a fair amount of complexity added to the macro with this
change, but on the other hand, it also succeeds in generating only the
necessary code. You win some, you lose some.