test(test2): Add UI tests for macro error output
The tests use the `trybuild` crate to check that compilation
fails for the incorrect macro use. The test will also fail
if the error output changes.

These were added directly to the `tests` directory
(as opposed to being part of the `testsuite` subdirectory)
because the automatic check and overwrite functionality 
provided by the `trybuild` crate seemed to not work properly
when placed there.